\name{acm.file.checkdelete}
\alias{acm.file.checkdelete}
\title{acm.file.checkdelete}
\usage{
acm.file.checkdelete(filename)
}
\arguments{
  \item{filename}{the name of the file which is tested for
  its existence}
}
\value{
none
}
\description{
acm.file.checkdelete abort the program if the file with the
given filename already exists
}
\author{
XiaoFei Zhao \email{xiaofei.zhao@mail.mcgill.ca}
}

