\name{acm.file.exists}
\alias{acm.file.exists}
\title{acm.file.exists (the modified file.exists function in acm)}
\usage{
acm.file.exists(filename)
}
\arguments{
  \item{filename}{the name of the file which is tested for
  its existence}
}
\value{
none
}
\description{
acm.file.exists behaves similar to file.exists except that
the program terminates and shows the already existing file
if the file already exists
}
\author{
XiaoFei Zhao \email{xiaofei.zhao@mail.mcgill.ca}
}

