\docType{data}
\name{experiments.in}
\alias{experiments.in}
\title{Sample fingerprint file}
\description{
Sample fingerprint file
}
\author{
Andrea Benedetti \email{andrea.benedetti@mcgill.ca}
}
\references{
Salamon et. al (1998) Accommodating Error Analysis in
Comparison and Clustering of Molecular Fingerprints.
Emerging Infectious Diseases Vol. 4, No. 2, April-June 1998

Abasci LLC. JAMES v1.0 User Documentation. 2002.
}
\keyword{data}

